Move register multisig view
