+++ src/modules/wallet/components/registerMultisigForm/feedback.js
