+++ src/modules/wallet/components/registerMultisigForm/form.test.js
