+++ src/modules/wallet/components/registerMultisigForm/index.js
