+++ src/modules/wallet/components/registerMultisigForm/styles.css
