+++ src/modules/wallet/components/registerMultisigStatus/index.js
