+++ src/modules/wallet/components/registerMultisigStatus/status.test.js
