+++ src/modules/wallet/components/registerMultisigSummary/index.js
