+++ src/modules/wallet/components/registerMultisigSummary/styles.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   box-sizing: border-box;
