+++ src/modules/wallet/components/registerMultisigSummary/summary.test.js
