+++ src/modules/wallet/components/registerMultisigView/progressBar.js
