+++ src/modules/wallet/components/registerMultisigView/styles.css
