+++ src/modules/wallet/manager/registerMultisigManager.js
@@ -1,0 +1,4 @@
+import { withRouter } from 'react-router-dom';
+import RegisterMultisigView from '../components/registerMultisigView';
+
+export default withRouter(RegisterMultisigView);
